+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/controller/BuyController.java
@@ -28,11 +28,11 @@ public class BuyController {
     private final ProductServiceAware productService;
 
     @GetMapping
-    public ModelAndView buy(@SessionAttribute("cartProductsMap") Map<Integer, Integer> cartProductsMap,
+    public ModelAndView buy(@SessionAttribute("cartProductsMap") Map<Long, Integer> cartProductsMap,
                             @SessionAttribute("user") User user) throws NotFoundException {
         Map<Product, Integer> products = new HashMap<>();
 
-        for (Map.Entry<Integer, Integer> entry : cartProductsMap.entrySet()) {
+        for (Map.Entry<Long, Integer> entry : cartProductsMap.entrySet()) {
             products.put(productService.findById(entry.getKey()), entry.getValue());
         }
