+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/repository/ImageRepository.java
@@ -3,5 +3,5 @@ package by.tms.eshopspringboot.repository;
 import by.tms.eshopspringboot.entity.Image;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface ImageRepository extends ListCrudRepository<Image, Integer> {
+public interface ImageRepository extends ListCrudRepository<Image, Long> {
 }
